+++ packages/studio-plugin/tests/FileSystemManager.test.ts
@@ -1,6 +1,6 @@
 import { Project } from "ts-morph";
 import FileSystemManager from "../src/FileSystemManager";
-import ParsingOrchestrator from "../src/ParsingOrchestrator";
+import ParsingOrchestrator from "../src/orchestrators/ParsingOrchestrator";
 import getUserPaths from "../src/parsers/getUserPaths";
 import upath from "upath";
 import fs from "fs";
